Refactoring part 1.
