+++ luisa_compute/examples/polymorphism_advanced.rs
@@ -1,9 +1,9 @@
 use std::env::current_exe;
 use std::f32::consts::PI;
 
+use luisa::lang::*;
 use luisa::prelude::*;
-use luisa::Value;
-use luisa::{impl_polymorphic, lang::*};
+use luisa::{impl_polymorphic, Value};
 use luisa_compute as luisa;
 
 #[derive(Clone, Hash, PartialEq, Eq, Debug)]
