+++ luisa_compute/src/lang.rs
@@ -1,0 +1,529 @@
+use std::any::Any;
+use std::cell::{Cell, RefCell};
+use std::collections::HashSet;
+use std::ffi::CString;
+use std::fmt::Debug;
+use std::marker::PhantomData;
+use std::ops::{Deref, DerefMut};
+use std::rc::Rc;
+use std::sync::atomic::AtomicUsize;
+use std::sync::Arc;
+use std::{env, unreachable};
+
+use half::f16;
+
+use bumpalo::Bump;
+use indexmap::IndexMap;
+
+use crate::runtime::{AsyncShaderArtifact, CallableArgEncoder, ShaderArtifact};
+use crate::{rtx, ResourceTracker};
+
+pub use luisa_compute_ir as ir;
+
+use ir::context::register_type;
+use ir::ffi::CBoxedSlice;
+use ir::ir::{
+    new_node, ArrayType, BasicBlock, Binding, CallableModule, CallableModuleRef, Capture, Const,
+    CpuCustomOp, Func, Instruction, IrBuilder, KernelModule, Module, ModuleFlags, ModuleKind,
+    ModulePools, Node, NodeRef, PhiIncoming, StructType, SwitchCase, Type, UserNodeData,
+    INVALID_REF,
+};
+use ir::{CArc, Pooled, TypeOf};
+
+use luisa_compute_derive::__Aggregate;
+
+// use self::math::Uint3;
+// pub mod maybe_expr;
+pub mod debug;
+pub mod diff;
+pub mod flow;
+pub mod functions;
+pub mod index;
+pub mod kernel;
+// pub mod ops;
+pub mod poly;
+pub mod printer;
+pub mod swizzle;
+pub mod types;
+
+use types::{Expr, Value, Var};
+use ops::{VarTrait, CommonVarOp, VarCmpEq, VarCmp, IntVarTrait, FloatVarTrait, ScalarVarTrait, VectorVarTrait, MatrixVarTrait, ScalarOrVector, BuiltinVarTrait}l
+
+// REMOVE
+use flow::*;
+use functions::*;
+use index::*;
+use kernel::*;
+use types::core::*;
+use types::vector::*;
+
+#[allow(dead_code)]
+pub(crate) static KERNEL_ID: AtomicUsize = AtomicUsize::new(0);
+// prevent node being shared across kernels
+// TODO: replace NodeRef with SafeNodeRef
+#[derive(Clone, Copy, Debug)]
+pub(crate) struct SafeNodeRef {
+    #[allow(dead_code)]
+    pub(crate) node: NodeRef,
+    #[allow(dead_code)]
+    pub(crate) kernel_id: usize,
+}
+
+pub trait Aggregate: Sized {
+    fn to_vec_nodes(&self) -> Vec<NodeRef> {
+        let mut nodes = vec![];
+        Self::to_nodes(&self, &mut nodes);
+        nodes
+    }
+    fn from_vec_nodes(nodes: Vec<NodeRef>) -> Self {
+        let mut iter = nodes.into_iter();
+        let ret = Self::from_nodes(&mut iter);
+        assert!(iter.next().is_none());
+        ret
+    }
+    fn to_nodes(&self, nodes: &mut Vec<NodeRef>);
+    fn from_nodes<I: Iterator<Item = NodeRef>>(iter: &mut I) -> Self;
+}
+
+impl<T: Aggregate> Aggregate for Vec<T> {
+    fn to_nodes(&self, nodes: &mut Vec<NodeRef>) {
+        let len_node = new_user_node(__module_pools(), nodes.len());
+        nodes.push(len_node);
+        for item in self {
+            item.to_nodes(nodes);
+        }
+    }
+
+    fn from_nodes<I: Iterator<Item = NodeRef>>(iter: &mut I) -> Self {
+        let len_node = iter.next().unwrap();
+        let len = len_node.unwrap_user_data::<usize>();
+        let mut ret = Vec::with_capacity(*len);
+        for _ in 0..*len {
+            ret.push(T::from_nodes(iter));
+        }
+        ret
+    }
+}
+
+impl<T: Aggregate> Aggregate for RefCell<T> {
+    fn to_nodes(&self, nodes: &mut Vec<NodeRef>) {
+        self.borrow().to_nodes(nodes);
+    }
+
+    fn from_nodes<I: Iterator<Item = NodeRef>>(iter: &mut I) -> Self {
+        RefCell::new(T::from_nodes(iter))
+    }
+}
+impl<T: Aggregate + Copy> Aggregate for Cell<T> {
+    fn to_nodes(&self, nodes: &mut Vec<NodeRef>) {
+        self.get().to_nodes(nodes);
+    }
+
+    fn from_nodes<I: Iterator<Item = NodeRef>>(iter: &mut I) -> Self {
+        Cell::new(T::from_nodes(iter))
+    }
+}
+impl<T: Aggregate> Aggregate for Option<T> {
+    fn to_nodes(&self, nodes: &mut Vec<NodeRef>) {
+        match self {
+            Some(x) => {
+                let node = new_user_node(__module_pools(), 1);
+                nodes.push(node);
+                x.to_nodes(nodes);
+            }
+            None => {
+                let node = new_user_node(__module_pools(), 0);
+                nodes.push(node);
+            }
+        }
+    }
+
+    fn from_nodes<I: Iterator<Item = NodeRef>>(iter: &mut I) -> Self {
+        let node = iter.next().unwrap();
+        let tag = node.unwrap_user_data::<usize>();
+        match *tag {
+            0 => None,
+            1 => Some(T::from_nodes(iter)),
+            _ => unreachable!(),
+        }
+    }
+}
+
+pub trait StructInitiaizable: Value {
+    type Init: Into<Self::Expr>;
+}
+
+pub trait ToNode {
+    fn node(&self) -> NodeRef;
+}
+
+pub trait FromNode: ToNode {
+    fn from_node(node: NodeRef) -> Self;
+}
+
+fn _store<T1: Aggregate, T2: Aggregate>(var: &T1, value: &T2) {
+    let value_nodes = value.to_vec_nodes();
+    let self_nodes = var.to_vec_nodes();
+    assert_eq!(value_nodes.len(), self_nodes.len());
+    __current_scope(|b| {
+        for (value_node, self_node) in value_nodes.into_iter().zip(self_nodes.into_iter()) {
+            b.update(self_node, value_node);
+        }
+    })
+}
+
+#[inline(always)]
+pub fn __new_user_node<T: UserNodeData>(data: T) -> NodeRef {
+    use luisa_compute_ir::ir::new_user_node;
+    new_user_node(__module_pools(), data)
+}
+macro_rules! impl_aggregate_for_tuple {
+    ()=>{
+        impl Aggregate for () {
+            fn to_nodes(&self, _: &mut Vec<NodeRef>) {}
+            fn from_nodes<I: Iterator<Item = NodeRef>>(_: &mut I) -> Self{}
+        }
+    };
+    ($first:ident  $($rest:ident) *) => {
+        impl<$first:Aggregate, $($rest: Aggregate),*> Aggregate for ($first, $($rest,)*) {
+            #[allow(non_snake_case)]
+            fn to_nodes(&self, nodes: &mut Vec<NodeRef>) {
+                let ($first, $($rest,)*) = self;
+                $first.to_nodes(nodes);
+                $($rest.to_nodes(nodes);)*
+            }
+            #[allow(non_snake_case)]
+            fn from_nodes<I: Iterator<Item = NodeRef>>(iter: &mut I) -> Self {
+                let $first = Aggregate::from_nodes(iter);
+                $(let $rest = Aggregate::from_nodes(iter);)*
+                ($first, $($rest,)*)
+            }
+        }
+        impl_aggregate_for_tuple!($($rest)*);
+    };
+
+}
+impl_aggregate_for_tuple!(T0 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11 T12 T13 T14 T15);
+
+pub unsafe trait Mask: ToNode {}
+
+pub fn select<A: Aggregate>(mask: impl Mask, a: A, b: A) -> A {
+    let a_nodes = a.to_vec_nodes();
+    let b_nodes = b.to_vec_nodes();
+    assert_eq!(a_nodes.len(), b_nodes.len());
+    let mut ret = vec![];
+    __current_scope(|b| {
+        for (a_node, b_node) in a_nodes.into_iter().zip(b_nodes.into_iter()) {
+            assert_eq!(a_node.type_(), b_node.type_());
+            assert!(!a_node.is_local(), "cannot select local variables");
+            assert!(!b_node.is_local(), "cannot select local variables");
+            if a_node.is_user_data() || b_node.is_user_data() {
+                assert!(
+                    a_node.is_user_data() && b_node.is_user_data(),
+                    "cannot select user data and non-user data"
+                );
+                let a_data = a_node.get_user_data();
+                let b_data = b_node.get_user_data();
+                if a_data != b_data {
+                    panic!("cannot select different user data");
+                }
+                ret.push(a_node);
+            } else {
+                ret.push(b.call(
+                    Func::Select,
+                    &[mask.node(), a_node, b_node],
+                    a_node.type_().clone(),
+                ));
+            }
+        }
+    });
+    A::from_vec_nodes(ret)
+}
+
+impl ToNode for bool {
+    fn node(&self) -> NodeRef {
+        const_(*self).node()
+    }
+}
+
+unsafe impl Mask for bool {}
+
+unsafe impl Mask for Bool {}
+
+extern "C" fn _trampoline<T, F: FnMut(&mut T)>(data: *mut u8, args: *mut u8) {
+    unsafe {
+        let container = &*(data as *const ClosureContainer<T>);
+        let f = &container.f;
+        let args = &mut *(args as *mut T);
+        f(args);
+    }
+}
+
+extern "C" fn _drop<T>(data: *mut u8) {
+    unsafe {
+        let _ = Box::from_raw(data as *mut T);
+    }
+}
+
+pub(crate) struct Recorder {
+    pub(crate) scopes: Vec<IrBuilder>,
+    pub(crate) kernel_id: Option<usize>,
+    pub(crate) lock: bool,
+    pub(crate) captured_buffer: IndexMap<Binding, (usize, NodeRef, Binding, Arc<dyn Any>)>,
+    pub(crate) cpu_custom_ops: IndexMap<u64, (usize, CArc<CpuCustomOp>)>,
+    pub(crate) callables: IndexMap<u64, CallableModuleRef>,
+    pub(crate) shared: Vec<NodeRef>,
+    pub(crate) device: Option<Device>,
+    pub(crate) block_size: Option<[u32; 3]>,
+    pub(crate) building_kernel: bool,
+    pub(crate) pools: Option<CArc<ModulePools>>,
+    pub(crate) arena: Bump,
+}
+
+impl Recorder {
+    fn reset(&mut self) {
+        self.scopes.clear();
+        self.captured_buffer.clear();
+        self.cpu_custom_ops.clear();
+        self.callables.clear();
+        self.lock = false;
+        self.device = None;
+        self.block_size = None;
+        self.arena.reset();
+        self.shared.clear();
+        self.kernel_id = None;
+    }
+    pub(crate) fn new() -> Self {
+        Recorder {
+            scopes: vec![],
+            lock: false,
+            captured_buffer: IndexMap::new(),
+            cpu_custom_ops: IndexMap::new(),
+            callables: IndexMap::new(),
+            shared: vec![],
+            device: None,
+            block_size: None,
+            pools: None,
+            arena: Bump::new(),
+            building_kernel: false,
+            kernel_id: None,
+        }
+    }
+}
+thread_local! {
+    pub(crate) static RECORDER: RefCell<Recorder> = RefCell::new(Recorder::new());
+}
+
+// Don't call this function directly unless you know what you are doing
+pub fn __current_scope<F: FnOnce(&mut IrBuilder) -> R, R>(f: F) -> R {
+    RECORDER.with(|r| {
+        let mut r = r.borrow_mut();
+        assert!(r.lock, "__current_scope must be called within a kernel");
+        let s = &mut r.scopes;
+        f(s.last_mut().unwrap())
+    })
+}
+
+pub(crate) fn __invoke_callable(callable: &CallableModuleRef, args: &[NodeRef]) -> NodeRef {
+    RECORDER.with(|r| {
+        let mut r = r.borrow_mut();
+        let id = CArc::as_ptr(&callable.0) as u64;
+        if let Some(c) = r.callables.get(&id) {
+            assert_eq!(CArc::as_ptr(&c.0), CArc::as_ptr(&callable.0));
+        } else {
+            r.callables.insert(id, callable.clone());
+        }
+    });
+    __current_scope(|b| {
+        b.call(
+            Func::Callable(callable.clone()),
+            args,
+            callable.0.ret_type.clone(),
+        )
+    })
+}
+
+pub(crate) fn __check_node_type(a: NodeRef, b: NodeRef) -> bool {
+    if !ir::context::is_type_equal(a.type_(), b.type_()) {
+        return false;
+    }
+    match (a.get().instruction.as_ref(), b.get().instruction.as_ref()) {
+        (Instruction::Buffer, Instruction::Buffer) => true,
+        (Instruction::Texture2D, Instruction::Texture2D) => true,
+        (Instruction::Texture3D, Instruction::Texture3D) => true,
+        (Instruction::Bindless, Instruction::Bindless) => true,
+        (Instruction::Accel, Instruction::Accel) => true,
+        (Instruction::Uniform, Instruction::Uniform) => true,
+        (Instruction::Local { .. }, Instruction::Local { .. }) => true,
+        (Instruction::Argument { by_value: true }, _) => b.get().instruction.has_value(),
+        (Instruction::Argument { by_value: false }, _) => b.is_lvalue(),
+        _ => false,
+    }
+}
+
+pub(crate) fn __check_callable(callable: &CallableModuleRef, args: &[NodeRef]) -> bool {
+    assert_eq!(callable.0.args.len(), args.len());
+    for i in 0..args.len() {
+        if !__check_node_type(callable.0.args[i], args[i]) {
+            return false;
+        }
+    }
+    true
+}
+
+// Don't call this function directly unless you know what you are doing
+pub fn __pop_scope() -> Pooled<BasicBlock> {
+    RECORDER.with(|r| {
+        let mut r = r.borrow_mut();
+        let s = &mut r.scopes;
+        s.pop().unwrap().finish()
+    })
+}
+
+pub fn __module_pools() -> &'static CArc<ModulePools> {
+    RECORDER.with(|r| {
+        let r = r.borrow();
+        assert!(r.lock, "__module_pools must be called within a kernel");
+        let pool = r.pools.as_ref().unwrap();
+        unsafe { std::mem::transmute(pool) }
+    })
+}
+// pub fn __load<T: Value>(node: NodeRef) -> Expr<T> {
+//     __current_scope(|b| {
+//         let node = b.load(node);
+//         Expr::<T>::from_node(node)
+//     })
+// }
+// pub fn __store(var:NodeRef, value:NodeRef) {
+//     let inst = &var.get().instruction;
+// }
+
+pub fn __extract<T: Value>(node: NodeRef, index: usize) -> NodeRef {
+    let inst = &node.get().instruction;
+    __current_scope(|b| {
+        let i = b.const_(Const::Int32(index as i32));
+        let op = match inst.as_ref() {
+            Instruction::Local { .. } => Func::GetElementPtr,
+            Instruction::Argument { by_value } => {
+                if *by_value {
+                    Func::ExtractElement
+                } else {
+                    Func::GetElementPtr
+                }
+            }
+            Instruction::Call(f, args) => match f {
+                Func::AtomicRef => {
+                    let mut indices = args.to_vec();
+                    indices.push(i);
+                    return b.call(Func::AtomicRef, &indices, <T as TypeOf>::type_());
+                }
+                _ => Func::ExtractElement,
+            },
+            _ => Func::ExtractElement,
+        };
+        let node = b.call(op, &[node, i], <T as TypeOf>::type_());
+        node
+    })
+}
+
+pub fn __insert<T: Value>(node: NodeRef, index: usize, value: NodeRef) -> NodeRef {
+    let inst = &node.get().instruction;
+    __current_scope(|b| {
+        let i = b.const_(Const::Int32(index as i32));
+        let op = match inst.as_ref() {
+            Instruction::Local { .. } => panic!("Can't insert into local variable"),
+            _ => Func::InsertElement,
+        };
+        let node = b.call(op, &[node, value, i], <T as TypeOf>::type_());
+        node
+    })
+}
+
+pub fn __compose<T: Value>(nodes: &[NodeRef]) -> NodeRef {
+    let ty = <T as TypeOf>::type_();
+    match ty.as_ref() {
+        Type::Struct(st) => {
+            assert_eq!(st.fields.as_ref().len(), nodes.len());
+            __current_scope(|b| b.call(Func::Struct, nodes, <T as TypeOf>::type_()))
+        }
+        Type::Primitive(_) => panic!("Can't compose primitive type"),
+        Type::Vector(vt) => {
+            let length = vt.length;
+            let func = match length {
+                2 => Func::Vec2,
+                3 => Func::Vec3,
+                4 => Func::Vec4,
+                _ => panic!("Can't compose vector with length {}", length),
+            };
+            __current_scope(|b| b.call(func, nodes, <T as TypeOf>::type_()))
+        }
+        Type::Matrix(vt) => {
+            let length = vt.dimension;
+            let func = match length {
+                2 => Func::Mat2,
+                3 => Func::Mat3,
+                4 => Func::Mat4,
+                _ => panic!("Can't compose vector with length {}", length),
+            };
+            __current_scope(|b| b.call(func, nodes, <T as TypeOf>::type_()))
+        }
+        _ => todo!(),
+    }
+}
+#[macro_export]
+macro_rules! struct_ {
+    ($t:ty { $($it:ident : $value:expr), *  $(,)?}) =>{
+        {
+            type Init = <$t as $crate::lang::StructInitiaizable>::Init;
+            let init = Init { $($it : $value), *  };
+            type Expr = <$t as $crate::lang::Value>::Expr;
+            let e:Expr = init.into();
+            e
+        }
+    }
+}
+
+pub const fn packed_size<T: Value>() -> usize {
+    (std::mem::size_of::<T>() + 3) / 4
+}
+
+pub fn pack_to<E, B>(expr: E, buffer: &B, index: impl Into<Expr<u32>>)
+where
+    E: ExprProxy,
+    B: IndexWrite<Element = u32>,
+{
+    let index = index.into();
+    __current_scope(|b| {
+        b.call(
+            Func::Pack,
+            &[expr.node(), buffer.node(), index.node()],
+            Type::void(),
+        );
+    });
+}
+
+pub fn unpack_from<T>(
+    buffer: &impl IndexWrite<Element = u32>,
+    index: impl Into<Expr<u32>>,
+) -> Expr<T>
+where
+    T: Value,
+{
+    let index = index.into();
+    Expr::<T>::from_node(__current_scope(|b| {
+        b.call(
+            Func::Unpack,
+            &[buffer.node(), index.node()],
+            <T as TypeOf>::type_(),
+        )
+    }))
+}
+
+pub(crate) fn need_runtime_check() -> bool {
+    cfg!(debug_assertions)
+        || match env::var("LUISA_DEBUG") {
+            Ok(s) => s == "full" || s == "1",
+            Err(_) => false,
+        }
+        || __env_need_backtrace()
+}
