+++ luisa_compute/src/lang/printer.rs
@@ -1,5 +1,4 @@
-use crate::prelude::*;
-use crate::*;
+use super::*;
 use parking_lot::RwLock;
 use std::fmt::Debug;
 use std::sync::atomic::AtomicBool;
