+++ luisa_compute/src/lib.rs
@@ -1,12 +1,13 @@
 #![allow(unused_unsafe)]
 
+use std::any::Any;
 use std::backtrace::Backtrace;
 use std::path::Path;
-use std::{any::Any, sync::Arc};
+use std::sync::Arc;
 
 pub mod lang;
 pub mod resource;
-pub mod rtx;
+// pub mod rtx;
 pub mod runtime;
 
 pub use half::f16;
@@ -15,8 +16,9 @@ pub use luisa_compute_backend as backend;
 
 pub mod prelude {
     pub use crate::lang::poly::PolymorphicImpl;
-    pub use crate::lang::traits::VarTrait;
-    pub use crate::lang::traits::{CommonVarOp, FloatVarTrait, IntVarTrait, VarCmp, VarCmpEq};
+    pub use crate::lang::traits::{
+        CommonVarOp, FloatVarTrait, IntVarTrait, VarCmp, VarCmpEq, VarTrait,
+    };
     pub use crate::lang::{
         Aggregate, ExprProxy, FromNode, IndexRead, IndexWrite, KernelBuildFn, KernelParameter,
         KernelSignature, Mask, Value, VarProxy,
@@ -41,26 +43,21 @@ pub use api::{
     AccelBuildModificationFlags, AccelBuildRequest, AccelOption, AccelUsageHint, MeshType,
     PixelFormat, PixelStorage,
 };
-pub use glam;
-pub use lang::math;
 pub use lang::math::*;
-pub use lang::poly;
 pub use lang::poly::*;
 pub use lang::traits::*;
-pub use lang::*;
-pub use log;
-pub use luisa_compute_derive as derive;
+pub use lang::{math, poly, *};
 pub use luisa_compute_derive::*;
 pub use luisa_compute_ir::ir::UserNodeData;
 pub use resource::*;
 pub use runtime::*;
+pub use {glam, log, luisa_compute_derive as derive};
 
 pub mod macros {
     pub use crate::{
-        cpu_dbg, if_, impl_new_poly_array, impl_polymorphic, lc_assert, lc_dbg, lc_unreachable,
-        loop_, struct_, var, while_,
+        cpu_dbg, if_, impl_new_poly_array, impl_polymorphic, lc_assert, lc_dbg, lc_debug, lc_error,
+        lc_info, lc_log, lc_unreachable, lc_warn, loop_, struct_, var, while_,
     };
-    pub use crate::{lc_debug, lc_error, lc_info, lc_log, lc_warn};
 }
 
 use lazy_static::lazy_static;
