+++ rustfmt.toml
@@ -1,0 +1,2 @@
+ignore = ["luisa_compute_sys"]
+imports_granularity = "Module"
